Add option to delete script

+++ clean.sh
@@ -3,7 +3,11 @@
 
 rm -rf "$HOME/.bashrc" "$HOME/.zshrc" "$HOME/.profile"
 rm -rf "$HOME/.editorconfig"
-rm -rf "$HOME/scripts"
+
+read -r -p 'Delete scripts? This includes extra scripts that may not be saved ' yn
+if [[ "$yn" == 'y' ]]; then
+    rm -rf "$HOME/scripts"
+fi
 
 rm -rf "$HOME/.vim/plugged"
 rm -rf "$HOME/.vim"
